+++ src/pyPLUTO/imagestate.py
@@ -22,7 +22,7 @@ class ImageState:
     dictcol: dict[int, str] = field(default_factory=dict)
     fig: Figure | None = None
     figsize: list[float] = field(default_factory=lambda: [8.0, 5.0])
-    fontsize: int = 17
+    fontsize: float = 17
     LaTeX: bool | str = True
     legpar: list[list[float]] = field(default_factory=list)
     legpos: list[int | str | None] = field(default_factory=list)
